Parser missing STR_VERBOSE
I will preface this by saying that I know nothing about C++ code and have never coded in C++ before.

However, I have been running into issues with dsk2kover stopping on 11%, not continuing and eventually just exiting with no error. However, I can't look at the info in the Kover dataset such as the number of kmers; I recieve an error saying:

KeyError: "Unable to open object (object 'kmer sequences' doesn't exist)" 

When verbose is turned on for creating the dataset, it does not print out anything extra at this step (DSK2Kover). If I am understanding how this code works correctly, I believe there needs to be a new line added for DSK2Kover taking in a STR_VERBOSE parameter because right now the code mentions STR_VERBOSE, but it appears this is not introduced the same way that all the other STR arguments are. I am not sure if the suggestion I made solves the issue or not, but I think this is something that does need to be fixed as it is making it very hard to trouble shoot the code.